Added *SubControllers and refactored ProductController to delegate logic to them.
